pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "downloads-agent"
version = "0.1.0"
description = "Automated Downloads organizer for macOS"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "downloads-agent contributors"},
]
keywords = ["macos", "downloads", "organizer", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Desktop Environment :: File Managers",
    "Topic :: System :: Filesystems",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
]

[project.scripts]
downloads-agent = "downloads_agent.cli:main"

[tool.setuptools.package-data]
downloads_agent = ["data/default.yaml"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["tests"]
addopts = "-v"
